+++ apps/syncro/CHANGELOG.md
@@ -5,11 +5,19 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @neurodesk/webapp-components@0.4.3
+  - @neurodesk/greedy@0.3.20260920
+  - @neurodesk/synthsr@0.4.20260920
+  - @neurodesk/brain-extraction@0.1.5
+
+
+### Patch Changes
+
+- Updated dependencies
   - @neurodesk/webapp-components@0.4.2
   - @neurodesk/greedy@0.3.20260920
   - @neurodesk/synthsr@0.4.20260920
   - @neurodesk/brain-extraction@0.1.4
-
 
 ### Patch Changes